Create hash method
It is basically the sum of the character ascii codes, but we multiply by
some number to avoid collisions between keys with the same characters
but shuffled around. The number used is a prime number since this
reduces collisions further due to the fact that there are not many ways
to exactly divide prime numbers I think.