Change length method to reading an instance var
The instance variable is updated correctly using the has? method which
is O(1) time in the set and remove methods.

The plan is to use the length method for every set and remove call to
check if the buckets array needs to grow. The old length method was O(n)
time since it called entries (which loops over buckets) and returned the
length of the entries array. So this would ruin the O(1) time intended
for set and remove. By making the length method a simple look up of the
instance variable, it is now O(1) which solves that problem. Also the
checks to update the length are also O(1), so now O(1) time is achieved
on set and remove.