Create initial hash set code (buggy)
Created a hash map instance variable in hash set to keep it simple.

For hash map we cannot rely on get() for has?() since get() returns nil
for both when the value is nil and when no key is found. This is a
problem for both hash map and hash set since I want the user to be able
to add nil values! I can see no other way other than to add a new method
to linked list, which will be the next commit.

I am ok with the behavior of get returning nil if no key is found since
something always has to be returned and nil makes the most sense. It
would be something to keep in mind for the user though! That is, if nil
is returned it either means that there is no key or the value is nil.
The user would then have to check using the correctly working has?()
for which one it is.